@@ -23,13 +23,13 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  charts_flutter: ^0.9.0
+  charts_flutter: ^0.10.0
   cupertino_icons: ^1.0.0
   date_range_picker: ^1.0.6
   esc_pos_bluetooth: ^0.2.8
   flutter_bloc: ^6.1.1
   introduction_screen: ^2.0.0
-  shared_preferences: ^0.5.12+4
+  shared_preferences: ^2.0.5
   sqflite: ^1.3.2+1
 
 dev_dependencies:
@@ -47,7 +47,7 @@ flutter:
   uses-material-design: true
   # To add assets to your application, add an assets section, like this:
   assets:
-    -  assets/img/
+    - assets/img/
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
   # For details regarding adding assets from package dependencies, see
